Move ServiceAttribute to new namespace
