+++ src/Miunie.Core/Services/ProfileService.cs
@@ -1,3 +1,4 @@
+using Miunie.Core.Attributes;
 using Miunie.Core.Discord;
 using Miunie.Core.Logging;
 using Miunie.Core.Providers;
